+++ src/main/java/be/dash/dashserver/core/domain/payment/PaymentClientApi.java
@@ -1,7 +1,6 @@
 package be.dash.dashserver.core.domain.payment;
 
-import be.dash.dashserver.api.core.lesson.dto.PaymentRequest;
-import be.dash.dashserver.core.domain.payment.dto.CancelReason;
+import be.dash.dashserver.core.domain.payment.dto.PaymentInformation;
 import be.dash.dashserver.core.domain.payment.dto.PaymentResult;
 
 public interface PaymentClientApi {
